Add application name on the top the screens Enhance the mvvm model to be consistent while rebuilding the screens
