+++ app/src/main/java/com/example/e_store/features/home/view/Home.kt
@@ -13,29 +13,10 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.e_store.features.home.view_model.HomeViewModel
-import com.example.e_store.features.home.view_model.HomeViewModelFactory
-import com.example.e_store.utils.data_layer.EStoreRepositoryImpl
-import com.example.e_store.utils.data_layer.local.room.EStoreLocalDataSourceImpl
-import com.example.e_store.utils.data_layer.remote.EStoreRemoteDataSourceImpl
 
 @Composable
-fun HomeScreen() {
-     val repo by lazy {
-        EStoreRepositoryImpl.getInstance(
-            EStoreRemoteDataSourceImpl.getInstance(),
-            EStoreLocalDataSourceImpl()
-        )
-    }
-
-    val homeViewModelFactory by lazy {
-       HomeViewModelFactory(repo)
-    }
-
-
-    val viewModel: HomeViewModel = viewModel(factory = homeViewModelFactory)
-
+fun HomeScreen(viewModel: HomeViewModel) {
     val uiState by viewModel.brands.collectAsStateWithLifecycle()
     val context = LocalContext.current
