+++ app/src/main/java/com/example/e_store/features/landing/view/LandingScreen.kt
@@ -9,11 +9,28 @@ import androidx.compose.material3.Surface
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.e_store.features.home.view_model.HomeViewModelFactory
 import com.example.e_store.features.landing.components.SplashLottie
 import com.example.e_store.features.main_home.view.MainHomeScreen
+import com.example.e_store.utils.data_layer.EStoreRepositoryImpl
+import com.example.e_store.utils.data_layer.local.room.EStoreLocalDataSourceImpl
+import com.example.e_store.utils.data_layer.remote.EStoreRemoteDataSourceImpl
 import com.example.e_store.utils.navigation.Screen
 
 class LandingScreen : ComponentActivity() {
+
+    private val repo by lazy {
+        EStoreRepositoryImpl.getInstance(
+            EStoreRemoteDataSourceImpl.getInstance(),
+            EStoreLocalDataSourceImpl()
+        )
+    }
+
+    val homeViewModelFactory by lazy {
+        HomeViewModelFactory(repo)
+    }
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
@@ -22,7 +39,9 @@ class LandingScreen : ComponentActivity() {
                     val navController = rememberNavController()
                     NavHost(navController, startDestination = Screen.Splash.route) {
                         composable(Screen.Splash.route) { SplashLottie(navController) }
-                        composable(Screen.Home.route) { MainHomeScreen() }
+                        composable(Screen.Home.route) { MainHomeScreen(
+                            homeViewModelFactory
+                        ) }
 
                     }
                 }
