+++ app/src/main/java/com/example/e_store/utils/navigation/Screen.kt
@@ -9,6 +9,7 @@ import com.example.e_store.R
 import com.example.e_store.features.categories.view.CategoriesScreen
 import com.example.e_store.features.home.view.HomeScreen
 import com.example.e_store.features.home.view_model.HomeViewModel
+import com.example.e_store.features.home.view_model.HomeViewModelFactory
 import com.example.e_store.features.profile.view.ProfileScreen
 import com.example.e_store.features.shopping_cart.view.ShoppingCartScreen
 import com.example.e_store.utils.constants.NavigationKeys
@@ -23,10 +24,11 @@ sealed class Screen(val route: String, val title: String, val icon: Int) {
     object Profile : Screen(NavigationKeys.PROFILE_ROUTE, "Profile", R.drawable.ic_person)
 }
     @Composable
-    fun AppNavigation(navController: NavHostController) {
+    fun AppNavigation(navController: NavHostController, homeViewModelFactory: HomeViewModelFactory) {
         NavHost(navController, startDestination = Screen.Home.route) {
             composable(Screen.Home.route) {
-                HomeScreen()
+                val viewModel: HomeViewModel = viewModel(factory = homeViewModelFactory)
+                HomeScreen(viewModel)
             }
             composable(Screen.Categories.route) { CategoriesScreen() }
             composable(Screen.Cart.route) { ShoppingCartScreen() }
